Move complete migration classes to demo module.
