+++ demo-simple/build.gradle
@@ -37,6 +37,8 @@ dependencies {
     implementation 'com.android.support:appcompat-v7:27.1.1'
     implementation project(path: ':library')
     testImplementation 'junit:junit:4.12'
+    testImplementation 'org.mockito:mockito-core:2.16.0'
+    testImplementation 'com.google.truth:truth:0.39'
     androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
