+++ demo-simple/src/test/java/com/novoda/downloadmanager/demo/migration/StubCursor.java
@@ -1,4 +1,4 @@
-package com.novoda.downloadmanager;
+package com.novoda.downloadmanager.demo.migration;
 
 import android.content.ContentResolver;
 import android.database.CharArrayBuffer;
